requirements.txt
openai==0.28
tiktoken 
chromadb==0.3.29
langchain==0.0.249
#pysqlite3-binary
flask
GitPython
python-dotenv
gunicorn
-e .